+++ js/components/coronavirus.js
@@ -1,7 +1,7 @@
 /* eslint-disable no-prototype-builtins */
 /* global Dashticz moment settings _CORS_PATH config language time blocks graph Chart _TEMP_SYMBOL getWeekNumber*/
-var api = 'https://coronavirus-tracker-api.herokuapp.com/v2/';
-var flagUrl = 'https://raw.githubusercontent.com/clinkadink/country-flags/master/png100px/';
+var api = _CORS_PATH + 'https://cvtapi.nl/v2/';
+var flagUrl = _CORS_PATH + 'https://raw.githubusercontent.com/clinkadink/country-flags/master/png100px/';
 var nf = Intl.NumberFormat();
 
 var DT_coronavirus = {